refactor(xtask): separate gates data types into submodule (#0000)
